S1: Extract config into ry-config library crate
Move Config, EnvironmentConfig, Excludes, ConfigError, Baseline,
load_baseline, write_baseline_file, subtract_baseline, build_filter,
and diagnostic_path from ry-cli into a new crates/ry-config library.

ry-cli re-imports via . ry-lsp declares
ry-config as a dependency (used in S2).

- ry check output is byte-identical (all 428 tests pass).
- Config tests move with the code and pass unchanged.
- dist-workspace.toml untouched (only ry-cli is published).
- cargo clippy --workspace -- -D warnings passes.
