+++ crates/ry-cli/Cargo.toml
@@ -19,6 +19,7 @@ path = "src/main.rs"
 ry-core = { path = "../ry-core" }
 ry-checker = { path = "../ry-checker" }
 ry-typeshed = { path = "../ry-typeshed" }
+ry-config = { path = "../ry-config" }
 ry-lsp = { path = "../ry-lsp" }
 clap = { workspace = true }
 clap_complete = { workspace = true }
