+++ crates/ry-lsp/Cargo.toml
@@ -11,6 +11,7 @@ authors.workspace = true
 ry-core = { path = "../ry-core" }
 ry-checker = { path = "../ry-checker" }
 ry-typeshed = { path = "../ry-typeshed" }
+ry-config = { path = "../ry-config" }
 tower-lsp = "0.20"
 tokio = { workspace = true, features = ["rt", "io-std", "sync", "time"] }
 tracing = { workspace = true }
